move specialty render helpers to helper
